Bank: rename package to it.polito.wa2.bankservicemock
